Update CLAUDE.md
This is mainly a big reduction in verbosity because I was observing Claude missing a lot of instructions - I think we preserved the essential parts but unfortunately with LLMs only time will tell if this is better or not. The main thing I'm expecting to see an improvement in is compliance with the Post-Change Checklist (or at least having to initiate it myself as much).