Fix HTML link format not working (resolves #59)
HTML formatting wasn't working due to an issue interacting with the clipboard and trying to format HTML using a rich text MIME type. HTML is instead now copied as a literal string which was always the original intent.

Note: RTF is still copied using that rich text format, so I have an item to look into that separately.